refactor: initialize room data and distribute component initialize existing room data when importing new room data distribute selector component
