+++ frontend/src/pages/Detail/Rooms.tsx
@@ -1,22 +1,23 @@
 import QueryString, { ParsedQs } from 'qs';
 import CheckInOut from 'components/CheckInOut/CheckInOut';
 import Room from 'components/Room/Room';
-import { useEffect, useState } from 'react';
+import { memo, useCallback, useEffect, useState } from 'react';
 import { getAllRoomList } from 'src/utils/requests';
-import { Buttons, SelectBtn, Selected, NotFoundRooms, SelectCartBtn } from './Rooms.style';
+import { NotFoundRooms } from './Rooms.style';
 import { addDays } from 'date-fns';
-import { Link, useLocation, useParams } from 'react-router-dom';
+import { useLocation, useParams } from 'react-router-dom';
 import Spinner from 'components/Spinner/Spinner';
 import { getReservedRooms } from 'src/utils/reservations';
 import { DetailRoomProps, RoomProps } from 'components/Room/Room.types';
 import changeDateFormatToIsoSTring from '../../utils/dateToISOString';
+import SelectBar from 'components/Room/Selector/SelectBar';
 
 const Rooms = () => {
   const { id } = useParams<string>();
   const location = useLocation();
   const { checkIn, checkOut }: ParsedQs = QueryString.parse(location.search, { ignoreQueryPrefix: true });
 
-  const [hotelId, setHotelId] = useState<string>(id);
+  const [hotelId] = useState<string>(id);
 
   const [rooms, setRooms] = useState<DetailRoomProps[] | undefined>([]);
   const [selectedRoom, setSelectedRoom] = useState<RoomProps>();
@@ -25,17 +26,23 @@ const Rooms = () => {
   const [startDate, setStartDate] = useState<Date>(new Date());
   const [endDate, setEndDate] = useState<Date>(addDays(new Date(), 1));
 
+  const setCheckInDate = useCallback((date:Date) => setStartDate(date), [startDate]);
+  const setCheckOutDate = useCallback((date:Date) => setEndDate(date), [endDate]);
+
+  const setSelector = useCallback((room:RoomProps)=>setSelectedRoom(room), [selectedRoom]);
+
   useEffect(() => {
     if (checkIn) setStartDate(new Date(checkIn));
     if (checkOut) setEndDate(new Date(checkOut));
   }, []);
 
-  useEffect(() => {
+  const setSessionStorage = useCallback(() => {
     window.sessionStorage.setItem('SELECTED_ROOM', JSON.stringify({ ...selectedRoom, startDate, endDate }));
   }, [selectedRoom]);
 
   useEffect(() => {
     const requestRooms = async () => {
+      setRooms([]);
       setIsLoaded(true);
       const checkIn = changeDateFormatToIsoSTring(startDate);
       const checkOut = changeDateFormatToIsoSTring(endDate);
@@ -56,37 +63,16 @@ const Rooms = () => {
 
   return (
     <>
-      <CheckInOut startDate={startDate} setStartDate={setStartDate} endDate={endDate} setEndDate={setEndDate} />
+      <CheckInOut startDate={startDate} setCheckInDate={setCheckInDate} endDate={endDate} setCheckOutDate={setCheckOutDate} />
       {isLoaded && <Spinner />}
       {rooms.length ? (
         <>
           <ul>
             {rooms.map((room, index) => (
-              <Room key={index} room={room} setSelectedRoom={setSelectedRoom} />
+              <Room key={index} room={room} setSelector={setSelector} />
             ))}
           </ul>
-          <Buttons>
-            {selectedRoom && selectedRoom.name ? (
-              <>
-                <Selected>
-                  현재 선택된 객실 : <span>{selectedRoom.name}</span>
-                </Selected>
-                <div>
-                  <Link to={'/cart'}>
-                    <SelectCartBtn>장바구니 담기</SelectCartBtn>
-                  </Link>
-                  <Link to={`/reservation/${hotelId}`}>
-                    <SelectBtn>예약하기</SelectBtn>
-                  </Link>
-                </div>
-              </>
-            ) : (
-              <div>
-                <SelectCartBtn disabled>장바구니 담기</SelectCartBtn>
-                <SelectBtn disabled>예약하기</SelectBtn>
-              </div>
-            )}
-          </Buttons>
+          <SelectBar selectedRoom={selectedRoom} hotelId={hotelId} setSessionStorage={setSessionStorage} />
         </>
       ) : (
         !isLoaded && (
@@ -100,4 +86,4 @@ const Rooms = () => {
   );
 };
 
-export default Rooms;
+export default memo(Rooms);
